put dedupVendoredDeps into VendoredDependency, and use it in ArchiveUploader
