+++ src/App/Fossa/LicenseScan.hs
@@ -8,7 +8,7 @@ import App.Fossa.Config.LicenseScan (
   mkSubCommand,
  )
 import App.Fossa.EmbeddedBinary (withThemisAndIndex)
-import App.Fossa.LicenseScanner (dedupVendoredDeps, scanVendoredDep)
+import App.Fossa.LicenseScanner (scanVendoredDep)
 import App.Fossa.ManualDeps (
   ManualDependencies (vendoredDependencies),
   findFossaDepsFile,
@@ -18,6 +18,7 @@ import App.Fossa.RunThemis (execRawThemis)
 import App.Fossa.Subcommand (SubCommand)
 import App.Fossa.VendoredDependency (
   VendoredDependency,
+  dedupVendoredDeps,
  )
 import App.Types (BaseDir (BaseDir))
 import Control.Carrier.StickyLogger (
