+++ src/App/Fossa/LicenseScanner.hs
@@ -4,7 +4,6 @@
 module App.Fossa.LicenseScanner (
   licenseScanSourceUnit,
   combineLicenseUnits,
-  dedupVendoredDeps,
   scanVendoredDep,
 ) where
 
@@ -16,12 +15,11 @@ import App.Fossa.VendoredDependency (
   VendoredDependency (..),
   arcToLocator,
   compressFile,
-  duplicateFailureBundle,
-  duplicateNames,
+  dedupVendoredDeps,
   hashFile,
  )
 import Control.Carrier.Finally (Finally, runFinally)
-import Control.Effect.Diagnostics (Diagnostics, ToDiagnostic (renderDiagnostic), context, fatal, fatalText, fromMaybe, recover)
+import Control.Effect.Diagnostics (Diagnostics, ToDiagnostic (renderDiagnostic), context, fatal, fromMaybe, recover)
 import Control.Effect.FossaApiClient (
   FossaApiClient,
   PackageRevision (..),
@@ -333,14 +331,3 @@ licenseScanSourceUnit baseDir vendoredDeps = do
 
     includeOrgId :: OrgId -> Archive -> Archive
     includeOrgId orgId arc = arc{archiveName = showT orgId <> "/" <> archiveName arc}
-
-dedupVendoredDeps :: (Has Diagnostics sig m) => NonEmpty VendoredDependency -> m (NonEmpty VendoredDependency)
-dedupVendoredDeps vdeps = do
-  -- Users with many instances of vendored dependencies may accidentally have complete duplicates. Remove them.
-  let uniqDeps = NE.nub vdeps
-  let duplicates = duplicateNames uniqDeps
-  case duplicates of
-    [] -> pure uniqDeps
-    -- However, users may also have vendored dependencies that have duplicate names but are not complete duplicates.
-    -- These aren't valid and can't be automatically handled, so fail the scan with them.
-    dups -> fatalText $ duplicateFailureBundle dups
